+++ apps/common/src/main/scala/org/lfdecentralizedtrust/splice/environment/MediatorAdminConnection.scala
@@ -8,11 +8,13 @@ import com.digitalasset.canton.admin.api.client.commands.{
   GrpcAdminCommand,
   MediatorAdminCommands,
   MediatorAdministrationCommands,
+  PruningSchedulerCommands,
   SequencerConnectionAdminCommands,
 }
 import com.digitalasset.canton.admin.api.client.data.{MediatorStatus, NodeStatus}
 import com.digitalasset.canton.config.{ApiLoggingConfig, ClientConfig}
 import com.digitalasset.canton.config.RequireTypes.{NonNegativeInt, PositiveInt}
+import com.digitalasset.canton.mediator.admin.v30.MediatorAdministrationServiceGrpc
 import com.digitalasset.canton.logging.NamedLoggerFactory
 import com.digitalasset.canton.sequencing.{
   SequencerConnection,
@@ -44,10 +46,24 @@ class MediatorAdminConnection(
       grpcClientMetrics,
       retryProvider,
     )
-    with StatusAdminConnection {
+    with StatusAdminConnection
+    with PruningAdminConnection {
 
   override val serviceName = "Canton Mediator Admin API"
 
+  override val pruningCommands: PruningSchedulerCommands[
+    MediatorAdministrationServiceGrpc.MediatorAdministrationServiceStub
+  ] = new PruningSchedulerCommands[
+    MediatorAdministrationServiceGrpc.MediatorAdministrationServiceStub
+  ](
+    MediatorAdministrationServiceGrpc.stub,
+    _.setSchedule(_),
+    _.clearSchedule(_),
+    _.setCron(_),
+    _.setMaxDuration(_),
+    _.setRetention(_),
+    _.getSchedule(_),
+  )
   override type Status = MediatorStatus
 
   override protected def getStatusRequest: GrpcAdminCommand[_, _, NodeStatus[MediatorStatus]] =